Added new craft files, edited others and did some coding this time too
so in this update I added all the stock size craft files for the new booster and edited the already existent craft files. All of them. I also did some changes to the telemetry with the booster and ship's orientation. I basically allowed them to rotate whichever way. This update is not full this is a part 1. There is a bug with the ship's orientation and 2 of the craft files aren't included so just be patient for part 2 of the update which will fix everything